Move SidebarActions to components
